Create OrderSystem.py
So this is so far the basic ordering system. I did try to keep it minimal however I needed a place holder for the avail_coils so that I could see how it would work, we may need to tweek this section a little so that it can read into our inventory.